updated mysql driver.
There is an older version of the github.com/go-sql-driver/mysql driver which sets `allowNativePasswords=false` by default. This breaks with hosted mysql service such as Azure MySQL, which requires native passwords to be allowed. Updating the driver to a version which contains this change will set the native passwords to be allowed by default.

Signed-off-by: Mike Lloyd <mike@reboot3times.org>